+++ public/scripts/dependencies/appwrite.js
@@ -2075,10 +2075,11 @@
              * function execution process will start asynchronously.
              *
              * @param {string} functionId
+             * @param {string} data 
              * @throws {Error}
              * @return {Promise}             
              */
-            createExecution: function(functionId) {
+            createExecution: function(functionId, data) {
                 if(functionId === undefined) {
                     throw new Error('Missing required parameter: "functionId"');
                 }
@@ -2086,6 +2087,10 @@
                 let path = '/functions/{functionId}/executions'.replace(new RegExp('{functionId}', 'g'), functionId);
 
                 let payload = {};
+
+                if (data) {
+                    payload['data'] = data;   
+                }
 
                 return http
                     .post(path, {